+++ src/app/src/components/AboutModal.jsx
@@ -40,52 +40,36 @@ const AboutModal = ({ isOpen, onClose }) => {
                     <Accordion allowToggle>
                         <StyledAccordionItem title='About the Survey on Gender Equality at Home'>
                             <StyledText>
-                                This survey was conducted by Facebook in
-                                consultation with the World Bank, UN Women,
-                                Equal Measures 2030, and Ladysmith.
+                                Achieving gender equality is one of society's
+                                greatest challenges. Too often, data that
+                                illuminates the lived realities of women and
+                                girls is missing, outdated, or unreliable.
+                                Through its Data for Good program to help
+                                researchers, nonprofits, and policymakers make
+                                informed and responsive decisions, Meta conducts
+                                a survey that covers topics about gender norms,
+                                unpaid and household care, access and agency,
+                                and COVID-19's impact on these areas.
                             </StyledText>
                             <StyledText>
-                                In July 2020, approximately half a million
-                                people in 208 countries, territories and islands
-                                were surveyed on Facebook to capture household
-                                gender dynamics during the COVID-19 pandemic.
-                                The data provides a global snapshot to help
-                                researchers, NGOs, global development
-                                institutions, and gender equality advocates
-                                identify and track trends and progress towards
-                                the UN Sustainable Development Goals (
+                                The survey has been fielded in 2020 and 2021
+                                over 200 countries. Public, aggregate datasets
+                                for the 2020 and 2021 surveys are available at
+                                the{' '}
                                 <StyledLink
-                                    href='https://sdgs.un.org/goals'
-                                    title='SDGs'
+                                    href='https://data.humdata.org/dataset/survey-on-gender-equality-at-home'
+                                    title='Humanitarian Data Exchange website'
                                 />
-                                ). In 2020, the COVID-19 pandemic provided a
-                                special impetus to examine these issues through
-                                the survey.
-                            </StyledText>{' '}
-                            <StyledText>
-                                The survey measures attitudes and beliefs about
-                                gender equality, participation in unpaid care
-                                and domestic work, resource allocation, and
-                                COVID-19’s impact on everyday lives. The data
-                                sheds light on differences within countries and
-                                among world regions regarding the extent to
-                                which people agree that men and women should
-                                have equal opportunities. The data also
-                                highlights gender and region-specific challenges
-                                of life during the pandemic, such as food
-                                insecurity, access to education and healthcare,
-                                and employment prospects.
-                            </StyledText>
-                            <StyledText>
-                                Contact{' '}
-                                <StyledLink
-                                    href='mailto:gendersurvey@fb.com'
-                                    title='gendersurvey@fb.com'
-                                />{' '}
-                                with any questions.
+                                , and de-identified, record-level response data
+                                to most questions may be available to research
+                                and nonprofit organizations for research
+                                purposes. Meta only shares data publicly once it
+                                is aggregated to the country or region level.
+                                Before sharing with Data for Good partners, Meta
+                                ensures that all survey data are de-identified.
                             </StyledText>
                         </StyledAccordionItem>
-                        <StyledAccordionItem title='How the survey contributes to Facebook’s mission and the UN Sustainable Development Goals'>
+                        <StyledAccordionItem title='How the survey contributes to Meta’s mission and the UN Sustainable Development Goals'>
                             <StyledText>
                                 This survey is part of{' '}
                                 <StyledLink
@@ -97,17 +81,17 @@ const AboutModal = ({ isOpen, onClose }) => {
                                     href='https://about.fb.com/news/2020/03/closing-the-gender-data-gap/'
                                     title='Project 17'
                                 />
-                                , two Facebook programs that use data to drive
+                                , two Meta programs that use data to drive
                                 progress on critical humanitarian issues. Data
                                 for Good provides tools built from
                                 privacy-protected data on Facebook’s platform,
                                 as well as tools developed using satellite
                                 imagery and other publicly available sources.
                                 Similarly, Project 17 emerged from a commitment
-                                Facebook made at the 2019 UN General Assembly to
+                                Meta made at the 2019 UN General Assembly to
                                 help partners use data to accelerate progress on
                                 the SDGs.
-                            </StyledText>{' '}
+                            </StyledText>
                             <StyledText>
                                 The Survey on Gender Equality at Home leverages
                                 Facebook’s global reach to fill important
@@ -137,30 +121,34 @@ const AboutModal = ({ isOpen, onClose }) => {
                                     href='https://facebookatunga.com/'
                                     title='here'
                                 />{' '}
-                                about how Facebook contributes to the SDGs.
+                                about how Meta contributes to the SDGs.
                             </StyledText>
                         </StyledAccordionItem>
                         <StyledAccordionItem title='Designing the survey'>
                             <StyledText>
-                                The survey was developed with inputs and
-                                collaboration from the World Bank, UN Women,
-                                Equal Measures 2030, and Ladysmith. Some of the
-                                questions have been borrowed from other surveys
-                                (e.g., face-to-face, telephone surveys, etc.),
-                                allowing for methodological comparisons. As
-                                such, the survey also generates insights useful
-                                to researchers undertaking alternative modes of
-                                data collection during COVID-19. The survey was
-                                also designed to avoid “survey fatigue,” where
-                                respondents begin to disengage from the survey
-                                content and responses become less reliable.
-                                Further details on the methodology are available{' '}
-                                <StyledLink
-                                    href='https://dataforgood.fb.com/wp-content/uploads/2020/09/Survey-on-Gender-Equality-at-Home-Report-1.pdf#page=60'
-                                    title='here'
-                                />
-                                .
-                            </StyledText>{' '}
+                                The survey’s questionnaires are designed in
+                                consultation with global gender equality
+                                experts. The 2020 wave was designed in
+                                consultation with the World Bank, UN Women,
+                                Equal Measures 2030, and Ladysmith. The 2021
+                                wave was designed in partnership with CARE,
+                                Ladysmith, the World Bank, and UNICEF.
+                            </StyledText>
+                            <StyledText>
+                                Some of the questions have been borrowed from
+                                other surveys (e.g., face-to-face, telephone
+                                surveys, etc.), allowing for methodological
+                                comparisons. As such, the survey also generates
+                                insights useful to researchers undertaking
+                                alternative modes of data collection during
+                                COVID-19.
+                            </StyledText>
+                            <StyledText>
+                                The questionnaires for each wave may vary. For
+                                example, about half of the questions in the 2021
+                                questionnaire are different from the 2020
+                                questionnaire.
+                            </StyledText>
                         </StyledAccordionItem>
                         <StyledAccordionItem title='Limitations of the data'>
                             <StyledText>
@@ -197,17 +185,18 @@ const AboutModal = ({ isOpen, onClose }) => {
                                     title='Humanitarian Data Exchange'
                                 />
                                 .
-                            </StyledText>{' '}
+                            </StyledText>
                             <StyledText>
                                 De-identified, record-level response data to
                                 most questions may be available to individuals
                                 and non-profit organizations for research
-                                purposes. Please email{' '}
+                                purposes. Visit the{' '}
                                 <StyledLink
-                                    href='mailto:gendersurvey@fb.com'
-                                    title='gendersurvey@fb.com'
+                                    href='https://dataforgood.facebook.com/dfg/tools/survey-on-gender-equality-at-home'
+                                    title='Data for Good at Meta website'
                                 />{' '}
-                                if you are interested in learning more.
+                                to learn more about how to request access to
+                                this microdata.
                             </StyledText>{' '}
                         </StyledAccordionItem>
                         <StyledAccordionItem title='Source code'>