+++ src/app/src/utils/index.js
@@ -18,10 +18,14 @@ export const isQuestionCode = key => {
     return key.includes('.');
 };
 
+const isCellEmpty = value => value == null || value.length === 0;
+
 export class DataIndexer {
-    constructor(year, geoMode, geographies, data) {
+    constructor(years, geoMode, geographies, data) {
         // Index the data by the current year and geography mode
-        this.data = data[geoMode][year];
+        // TODO: Enable multi-year question handling
+        this.year = years[0];
+        this.data = data[geoMode][years[0]];
         this.geographies = geographies;
         this.survey = CONFIG[geoMode].survey;
     }
@@ -57,13 +61,16 @@ export class DataIndexer {
     formatForViz(key, geo) {
         const resp = this.survey[key];
         if (!resp) {
-            return { key, geo };
+            return { key, geo, dataUnavailable: true };
         }
+
         const idx = resp.idx;
         const d = this.data.geographies[geo];
-        const c = d['Combined'][idx].toFixed(2);
-        const m = d['Male'][idx].toFixed(2);
-        const f = d['Female'][idx].toFixed(2);
+        const c = this.formatCell(d, 'Combined', idx);
+        const m = this.formatCell(d, 'Male', idx);
+        const f = this.formatCell(d, 'Female', idx);
+
+        const dataUnavailable = c == null && m == null && f == null;
 
         return {
             key: key,
@@ -72,7 +79,14 @@ export class DataIndexer {
             combined: c,
             female: f,
             male: m,
+            year: this.year,
+            dataUnavailable,
         };
+    }
+
+    formatCell(d, g, idx) {
+        const cell = d[g][idx];
+        return isCellEmpty(cell) ? null : cell.toFixed(2);
     }
 
     getAllResponsesForQ(qcode) {
@@ -89,6 +103,62 @@ export class DataIndexer {
         const { cat, ...rest } = response;
         return rest;
     }
+
+    isDataUnavailable(key) {
+        // There is no data available for the question
+        // for any selected geography in the current year
+        return this.geographies.every(
+            geo => this.formatForViz(key, geo).dataUnavailable
+        );
+    }
+
+    getGeoAvailability(key) {
+        // Returns an object showing the question's
+        // availability for each geography
+        return this.geographies.reduce(
+            (acc, geo) => ({
+                ...acc,
+                [geo]: this.formatForViz(key, geo).dataUnavailable,
+            }),
+            {}
+        );
+    }
 }
 
 export const formatQuery = str => str?.trim().toLowerCase() || '';
+
+export const calculateAvailableGeo = ({
+    years,
+    geoMode,
+    currentGeo,
+    data,
+    survey,
+}) => {
+    if (!currentGeo.length) return [];
+    return years.reduce((availableYears, year) => {
+        const indexer = new DataIndexer([year], geoMode, currentGeo, data);
+        const geoAvailability = Object.entries(survey).map(([key, question]) =>
+            indexer.getGeoAvailability(key)
+        );
+
+        return {
+            ...availableYears,
+            [year]: currentGeo.filter(geo =>
+                geoAvailability.some(ga => !ga[geo])
+            ),
+        };
+    }, {});
+};
+
+export const formatCurrentGeo = ({
+    currentGeo,
+    currentYears,
+    availableYearsGeography,
+}) =>
+    currentGeo
+        .filter(geo =>
+            currentYears.some(year =>
+                availableYearsGeography[year].includes(geo)
+            )
+        )
+        .join(', ');